pyproject.toml
[project]
name = "ai-cost-modeler"
version = "0.1.0"
description = "Model and compare AI inference costs before you build"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "tiktoken>=0.12.0",
    "click>=8.3.1",
    "rich>=14.3.3",
    "duckdb>=1.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
cost = "costmodel.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["costmodel"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
